Fixed CI failure for windows-benchmark
Fixes https://github.com/gfoidl/Base64/issues/20

+++ run-benchmarks.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
-
+#------------------------------------------------------------------------------
 set -e
+#------------------------------------------------------------------------------
+if [[ "$1" == "--install-sdk" ]]; then
+    echo "installing SDK $SDK_VERSION"
+    echo ""
 
-curl -o dotnet-install.sh https://dot.net/v1/dotnet-install.sh
-mkdir dotnet
-chmod u+x ./dotnet-install.sh
-./dotnet-install.sh --install-dir $(pwd)/dotnet -v $SDK_VERSION
-rm $(pwd)/dotnet-install.sh
-export PATH="$(pwd)/dotnet:$PATH"
+    curl -o dotnet-install.sh https://dot.net/v1/dotnet-install.sh
+    mkdir dotnet
+    chmod u+x ./dotnet-install.sh
+    ./dotnet-install.sh --install-dir $(pwd)/dotnet -v $SDK_VERSION
+    rm $(pwd)/dotnet-install.sh
+    export PATH="$(pwd)/dotnet:$PATH"
+
+    echo ""
+fi
 
 echo 'installed sdks:'
 dotnet --list-sdks
